pyproject.toml
[tool.poetry]
name = "diplom-1"
version = "0.1.0"
description = ""
authors = ["evgeniatimoshkina <tea302@yandex.ru>"]
readme = "README.md"
packages = [{include = "diplom_1"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
docker-compose = "^1.29.2"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.1"
social-auth-app-django = "^5.2.0"
django-filter = "^23.2"
pytest = "^7.3.1"
marshmallow-dataclass = "^8.5.14"
pytest-factory = "^0.0.1.dev0"
pytest-factoryboy = "^2.5.1"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
